buildRustCrate: revert 4c89619152a9084f104f883145298c8c89b0616d per issue 130309
Change buildRustCrate's codgen-units option from $NIX_BUILD_CORES to 1.

This is necessary because codegen-units is not a pure optimization, but substantially affects the resulting binary.
The current version of buildRustCrate gives different binaries for the same derivation depending on the `--cores` option, which results in failing `--check` if you've tried two different ones.

Any static value will do; I chose 1 because (a) it was what nixpkgs used before changing to a variable and (b) it results in the fastest running binary.
16 (the cargo default) would also have been a sensible choice, as would making buildRustCrate configurable.

See nixpkgs issue 130309 for more discussion.